# Cleanup predicate simplifier code