We currently have two code paths for predicate simplification. This is messy and confusing. This PR refactors this into a single code-path parametrized on whether we want to recurse into equation application or not.