@goodlyrottenapple 2023-11-15
```
kevm-master-449fbfa-master-8ecb99be22406efbdf151eb20a8d253d72a961f2-sam-cleanup-predicate-simplifier-compare
| Test                                      | sam-cleanup-predicate-simplifier time | master time | (sam-cleanup-predicate-simplifier/master) time
|-------------------------------------------|---------------------------------------|-------------|------------------------------------------------
| mcd/dsvalue-read-pass-summarize-spec.k    | 72.85                                 | 82.89       | 0.8788756182892989
| mcd/dsvalue-read-pass-spec.k              | 86.3                                  | 92.22       | 0.935805682064628
| benchmarks/storagevar02-nooverflow-spec.k | 64.74                                 | 68.03       | 0.951638982801705
| examples/solidity-code-spec.md            | 251.93                                | 262.3       | 0.9604651162790697
| mcd/flipper-tau-pass-spec.k               | 93.21                                 | 83.75       | 1.1129552238805969
| TOTAL                                     | 569.03                                | 589.19      | 0.9657835333254127
```
```
kontrol-v0.1.49-master-15a9fc41edcf6b22730f37e77a3e29acb2ae3601-sam-map-unification-improvements-compare
 Test                                                                                                                      | sam-map-unification-improvements time | master time | (sam-map-unification-improvements/master) time
---------------------------------------------------------------------------------------------------------------------------+---------------------------------------+-------------+------------------------------------------------
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[AddrTest.test_notBuiltinAddress_concrete()]               | 55.89                                 | 53.95       | 1.0359592215013902
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[PlainPrankTest.test_startPrankWithOrigin_true()]          | 211.98                                | 204.39      | 1.0371348891824455
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[ExpectRevertTest.testFail_expectRevert_empty()]           | 76.69                                 | 73.88       | 1.0380346507850569
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[ExpectRevertTest.test_expectRevert_message()]             | 151.49                                | 145.89      | 1.0383850846528209
 src/tests/integration/test_foundry_prove.py::test_foundry_fail[AssertTest.test_revert_branch(uint256,uint256)]            | 114.17                                | 109.12      | 1.0462793255131964
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[ArithmeticTest.test_max1(uint256,uint256)]                | 123.0                                 | 117.47      | 1.0470758491529752
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[LoopsTest.test_sum_10()]                                  | 96.54                                 | 91.73       | 1.052436498419274
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[EmitContractTest.testExpectEmitCheckEmitter()]            | 143.76                                | 135.58      | 1.060333382504794
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[AddrTest.test_addr_true()]                                | 66.52                                 | 62.15       | 1.0703137570394208
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[ExpectRevertTest.testFail_ExpectRevert_failAndSuccess()]  | 163.32                                | 151.16      | 1.0804445620534533
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[ExpectRevertTest.testFail_expectRevert_multipleReverts()] | 179.1                                 | 164.95      | 1.085783570779024
 src/tests/integration/test_foundry_prove.py::test_foundry_prove[BMCLoopsTest.test_countdown_concrete()]                   | 62.16                                 | 56.58       | 1.0986214209968186
 src/tests/integration/test_solc_to_k.py::test_solc_to_k[Empty.sol]                                                        | 55.9                                  | 50.21       | 1.1133240390360486
 src/tests/integration/test_solc_to_k.py::test_solc_to_k[ERC721.sol]                                                       | 56.53                                 | 48.85       | 1.1572159672466735
 TOTAL                                                                                                                     | 1557.05                               | 1465.91     | 1.062172984698924
```