@jberthold 2023-11-15
I quickly updated my branch from #352 to latest master, maybe we can combine the two lines of code? 
That seems to be the best way IMHO